refactor: 커스텀미션 status가 state를 가지고 가는 형태로 변경
